refactor module name aggregation
