+++ pyam/aggregation.py
